refactor: Consolidate handler files to eliminate duplication
Problem: Having separate 'smart_handlers.go' and 'advanced_handlers.go'
was confusing and implied artificial separation.

Solution: Merged into single 'handlers.go' with clear sections:

## File Structure (Before):
- main.go (240 lines) - Entry point
- smart_handlers.go (240 lines) - index, search, version, versions
- advanced_handlers.go (389 lines) - chat, query, diff, info, verify
Total: 869 lines across 3 files

## File Structure (After):
- main.go (140 lines) - Entry point, basic commands, usage
- handlers.go (600 lines) - ALL handlers organized by category
Total: 740 lines across 2 files ✅

## Organization in handlers.go:

Benefits:
✅ No duplicate imports
✅ Clear logical grouping
✅ Single source of truth
✅ Easier to navigate
✅ No confusion about which file to edit
