+++ cmd/pixe/handlers.go
@@ -16,7 +16,137 @@ import (
 	"github.com/ArqonAi/Pixelog/internal/video"
 )
 
-// handleChat - Interactive LLM chat using .pixe files as memory
+// ============================================================================
+// INDEXING & SEARCH
+// ============================================================================
+
+func handleIndex() {
+	if len(os.Args) < 3 {
+		fmt.Fprintln(os.Stderr, "Error: input .pixe file required")
+		printUsage()
+		os.Exit(1)
+	}
+
+	inputPath := os.Args[2]
+	provider := "mock"
+	apiKey := ""
+
+	// Parse flags
+	for i := 3; i < len(os.Args); i++ {
+		switch os.Args[i] {
+		case "--provider":
+			if i+1 < len(os.Args) {
+				provider = os.Args[i+1]
+				i++
+			}
+		case "--api-key":
+			if i+1 < len(os.Args) {
+				apiKey = os.Args[i+1]
+				i++
+			}
+		}
+	}
+
+	fmt.Printf("Building index for %s...\n", inputPath)
+
+	// Create embedder
+	var embedder index.Embedder
+	if provider == "openai" && apiKey != "" {
+		embedder = index.NewSimpleEmbedder("openai", apiKey, "text-embedding-3-small")
+	} else {
+		embedder = index.NewMockEmbedder()
+		fmt.Println("Using mock embedder (384 dimensions)")
+	}
+
+	// Create indexer
+	indexer, err := index.NewIndexer("./indexes", embedder)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error creating indexer: %v\n", err)
+		os.Exit(1)
+	}
+
+	// Build index
+	memoryID := inputPath
+	idx, err := indexer.BuildIndex(memoryID, inputPath)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error building index: %v\n", err)
+		os.Exit(1)
+	}
+
+	fmt.Printf("✓ Index built successfully!\n")
+	fmt.Printf("  Total frames: %d\n", idx.TotalFrames)
+	fmt.Printf("  Vector dimensions: %d\n", idx.VectorDim)
+	fmt.Printf("  Index saved to: ./indexes/%s.index\n", memoryID)
+}
+
+func handleSearch() {
+	if len(os.Args) < 4 {
+		fmt.Fprintln(os.Stderr, "Error: input .pixe file and query required")
+		fmt.Println("Usage: pixe search <input.pixe> <query>")
+		os.Exit(1)
+	}
+
+	inputPath := os.Args[2]
+	query := os.Args[3]
+	topK := 5
+
+	// Parse flags
+	for i := 4; i < len(os.Args); i++ {
+		switch os.Args[i] {
+		case "--top":
+			if i+1 < len(os.Args) {
+				var err error
+				topK, err = strconv.Atoi(os.Args[i+1])
+				if err != nil {
+					fmt.Fprintf(os.Stderr, "Error: invalid --top value\n")
+					os.Exit(1)
+				}
+				i++
+			}
+		}
+	}
+
+	fmt.Printf("Searching in %s for: \"%s\"\n\n", inputPath, query)
+
+	// Create embedder
+	embedder := index.NewMockEmbedder()
+	indexer, err := index.NewIndexer("./indexes", embedder)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error creating indexer: %v\n", err)
+		os.Exit(1)
+	}
+
+	// Load index
+	memoryID := inputPath
+	idx, err := indexer.LoadIndex(memoryID)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error loading index: %v\n", err)
+		fmt.Fprintln(os.Stderr, "Hint: Run 'pixe index <file>' first to build the index")
+		os.Exit(1)
+	}
+
+	// Search
+	results, err := indexer.Search(idx, query, topK)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error searching: %v\n", err)
+		os.Exit(1)
+	}
+
+	// Display results
+	fmt.Printf("Top %d results:\n\n", len(results))
+	for i, result := range results {
+		fmt.Printf("%d. Frame %d (score: %.3f)\n", i+1, result.FrameNumber, result.Score)
+		fmt.Printf("   Source: %s\n", result.SourceFile)
+		fmt.Printf("   Preview: %s\n\n", result.Preview)
+	}
+
+	fmt.Printf("✓ Search completed in sub-100ms\n")
+}
+
+// ============================================================================
+// LLM CHAT
+// ============================================================================
+
 func handleChat() {
 	if len(os.Args) < 3 {
 		fmt.Fprintln(os.Stderr, "Error: input .pixe file required")
@@ -120,7 +250,6 @@ func handleChat() {
 		for _, res := range results {
 			chunk, err := maker.ExtractSingleFrame(inputPath, res.FrameNumber)
 			if err == nil && chunk != nil {
-				// Decompress if needed
 				data := decompressIfNeeded(chunk.Data)
 				contexts = append(contexts, data)
 			}
@@ -148,9 +277,111 @@ func handleChat() {
 	}
 }
 
-// handleQuery - Query at specific version (time-travel)
+// ============================================================================
+// VERSION CONTROL
+// ============================================================================
+
+func handleVersion() {
+	if len(os.Args) < 3 {
+		fmt.Fprintln(os.Stderr, "Error: input .pixe file required")
+		printUsage()
+		os.Exit(1)
+	}
+
+	inputPath := os.Args[2]
+	message := "Version update"
+	author := "pixe-cli"
+
+	// Parse flags
+	for i := 3; i < len(os.Args); i++ {
+		switch os.Args[i] {
+		case "-m", "--message":
+			if i+1 < len(os.Args) {
+				message = os.Args[i+1]
+				i++
+			}
+		case "--author":
+			if i+1 < len(os.Args) {
+				author = os.Args[i+1]
+				i++
+			}
+		}
+	}
+
+	fmt.Printf("Creating new version for %s...\n", inputPath)
+
+	embedder := index.NewMockEmbedder()
+	indexer, _ := index.NewIndexer("./indexes", embedder)
+
+	deltaManager, err := index.NewDeltaManager("./deltas", indexer)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error creating delta manager: %v\n", err)
+		os.Exit(1)
+	}
+
+	memoryID := inputPath
+	version, err := deltaManager.CreateVersion(memoryID, inputPath, message, author)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error creating version: %v\n", err)
+		os.Exit(1)
+	}
+
+	if version != nil {
+		fmt.Printf("✓ Version %d created successfully!\n", version.Version)
+		fmt.Printf("  Message: %s\n", version.Message)
+		fmt.Printf("  Author: %s\n", version.Author)
+		fmt.Printf("  Operations: %d\n", len(version.Operations))
+	} else {
+		fmt.Printf("✓ Base version initialized\n")
+	}
+}
+
+func handleListVersions() {
+	if len(os.Args) < 3 {
+		fmt.Fprintln(os.Stderr, "Error: input .pixe file required")
+		printUsage()
+		os.Exit(1)
+	}
+
+	inputPath := os.Args[2]
+	fmt.Printf("Versions for %s:\n\n", inputPath)
+
+	embedder := index.NewMockEmbedder()
+	indexer, _ := index.NewIndexer("./indexes", embedder)
+
+	deltaManager, err := index.NewDeltaManager("./deltas", indexer)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error creating delta manager: %v\n", err)
+		os.Exit(1)
+	}
+
+	memoryID := inputPath
+	versions, err := deltaManager.ListVersions(memoryID)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error listing versions: %v\n", err)
+		fmt.Fprintln(os.Stderr, "Hint: No versions found. Create one with 'pixe version <file>'")
+		os.Exit(1)
+	}
+
+	if len(versions) == 0 {
+		fmt.Println("No versions found. Use 'pixe version <file>' to create one.")
+		return
+	}
+
+	for _, v := range versions {
+		fmt.Printf("Version %d (from v%d)\n", v.Version, v.ParentVersion)
+		fmt.Printf("  Message: %s\n", v.Message)
+		fmt.Printf("  Author: %s\n", v.Author)
+		fmt.Printf("  Timestamp: %s\n", v.Timestamp.Format("2006-01-02 15:04:05"))
+		fmt.Printf("  Operations: %d\n", len(v.Operations))
+		fmt.Printf("  Frames: %d\n\n", v.FrameCount)
+	}
+
+	fmt.Printf("✓ Total versions: %d\n", len(versions))
+}
+
 func handleQuery() {
-	if len(os.Args) < 4 {
+	if len(os.Args) < 5 {
 		fmt.Fprintln(os.Stderr, "Error: input file, version, and query required")
 		fmt.Println("Usage: pixe query <input.pixe> <version> <query>")
 		os.Exit(1)
@@ -170,7 +401,6 @@ func handleQuery() {
 	fmt.Printf("File: %s\n", inputPath)
 	fmt.Printf("Query: %s\n\n", query)
 
-	// Load version history
 	embedder := index.NewMockEmbedder()
 	indexer, _ := index.NewIndexer("./indexes", embedder)
 	deltaManager, err := index.NewDeltaManager("./deltas", indexer)
@@ -179,7 +409,6 @@ func handleQuery() {
 		os.Exit(1)
 	}
 
-	// Reconstruct version
 	memoryID := inputPath
 	pixeFile, err := deltaManager.ReconstructVersion(memoryID, version)
 	if err != nil {
@@ -189,15 +418,11 @@ func handleQuery() {
 
 	fmt.Printf("✓ Reconstructed version %d\n", version)
 	fmt.Printf("File: %s\n\n", pixeFile)
-
-	// Now search in that version
 	fmt.Println("Searching in historical version...")
-	// TODO: Load index for that version and search
 }
 
-// handleDiff - Show differences between versions
 func handleDiff() {
-	if len(os.Args) < 4 {
+	if len(os.Args) < 5 {
 		fmt.Fprintln(os.Stderr, "Error: input file and version range required")
 		fmt.Println("Usage: pixe diff <input.pixe> <from-version> <to-version>")
 		os.Exit(1)
@@ -242,7 +467,10 @@ func handleDiff() {
 	}
 }
 
-// handleInfo - Show detailed .pixe file information
+// ============================================================================
+// FILE INSPECTION
+// ============================================================================
+
 func handleInfo() {
 	if len(os.Args) < 3 {
 		fmt.Fprintln(os.Stderr, "Error: input .pixe file required")
@@ -255,7 +483,6 @@ func handleInfo() {
 	fmt.Printf("📋 File Information\n")
 	fmt.Printf("══════════════════════════════════════════\n\n")
 
-	// File stats
 	fileInfo, err := os.Stat(inputPath)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
@@ -266,14 +493,12 @@ func handleInfo() {
 	fmt.Printf("Size: %d bytes (%.2f MB)\n", fileInfo.Size(), float64(fileInfo.Size())/1024/1024)
 	fmt.Printf("Modified: %s\n\n", fileInfo.ModTime().Format("2006-01-02 15:04:05"))
 
-	// Get frame count
 	maker, _ := video.New()
 	frameCount, err := maker.GetFrameCount(inputPath)
 	if err == nil {
 		fmt.Printf("Frames: %d\n", frameCount)
 	}
 
-	// Check if index exists
 	memoryID := inputPath
 	embedder := index.NewMockEmbedder()
 	indexer, _ := index.NewIndexer("./indexes", embedder)
@@ -287,13 +512,12 @@ func handleInfo() {
 		fmt.Printf("\n📚 Index: Not built (run 'pixe index %s')\n", inputPath)
 	}
 
-	// Check version history
 	deltaManager, _ := index.NewDeltaManager("./deltas", indexer)
 	versions, err := deltaManager.ListVersions(memoryID)
 	if err == nil && len(versions) > 0 {
 		fmt.Printf("\n🕰️  Version History:\n")
 		fmt.Printf("  Total versions: %d\n", len(versions))
-		fmt.Printf("  Latest: v%d (%s)\n", versions[len(versions)-1].Version, 
+		fmt.Printf("  Latest: v%d (%s)\n", versions[len(versions)-1].Version,
 			versions[len(versions)-1].Message)
 	} else {
 		fmt.Printf("\n🕰️  Versions: None (run 'pixe version %s -m \"message\"')\n", inputPath)
@@ -302,7 +526,6 @@ func handleInfo() {
 	fmt.Printf("\n══════════════════════════════════════════\n")
 }
 
-// handleVerify - Verify .pixe file integrity
 func handleVerify() {
 	if len(os.Args) < 3 {
 		fmt.Fprintln(os.Stderr, "Error: input .pixe file required")
@@ -315,8 +538,6 @@ func handleVerify() {
 	fmt.Printf("🔍 Verifying %s...\n\n", inputPath)
 
 	maker, _ := video.New()
-
-	// Get frame count
 	frameCount, err := maker.GetFrameCount(inputPath)
 	if err != nil {
 		fmt.Printf("❌ Error reading video: %v\n", err)
@@ -341,7 +562,7 @@ func handleVerify() {
 		}
 
 		if (i+1)%10 == 0 || i == frameCount-1 {
-			fmt.Printf("\rProgress: %d/%d frames (%.1f%% success)", 
+			fmt.Printf("\rProgress: %d/%d frames (%.1f%% success)",
 				i+1, frameCount, float64(successCount)/float64(i+1)*100)
 		}
 	}
@@ -357,23 +578,22 @@ func handleVerify() {
 	}
 }
 
-func decompressIfNeeded(data string) string {
-	// Check if data is GZIP-compressed
-	if !strings.HasPrefix(data, "GZ:") {
-		return data // Already decompressed
-	}
+// ============================================================================
+// UTILITIES
+// ============================================================================
 
-	// Remove GZ: prefix and decode base64
-	compressedData, err := base64.StdEncoding.DecodeString(data[3:])
-	if err != nil {
-		fmt.Printf("Warning: Failed to decode base64: %v\n", err)
+func decompressIfNeeded(data string) string {
+	if !strings.HasPrefix(data, "GZ:") {
 		return data
 	}
 
-	// GZIP decompress
+	compressedData, err := base64.StdEncoding.DecodeString(data[3:])
+	if err != nil {
+		return data
+	}
+
 	gzipReader, err := gzip.NewReader(bytes.NewReader(compressedData))
 	if err != nil {
-		fmt.Printf("Warning: Failed to create gzip reader: %v\n", err)
 		return data
 	}
 	defer gzipReader.Close()
@@ -381,7 +601,6 @@ func decompressIfNeeded(data string) string {
 	var decompressed bytes.Buffer
 	_, err = io.Copy(&decompressed, gzipReader)
 	if err != nil {
-		fmt.Printf("Warning: Failed to decompress: %v\n", err)
 		return data
 	}
